requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
captcha==0.4
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-paypal==2.0
django-ranged-response==0.2.0
django-robokassa-payments==1.3.3
django-simple-captcha==0.5.17
django-timezone-field==4.2.3
djangorestframework==3.13.1
docker==5.0.3
idna==3.3
kombu==5.2.4
lxml==4.8.0
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
redis-server==6.0.9
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.1
